refactor: move out of utils dir
