+++ packages/portal/src/plugins/europeana/entity-management.js
@@ -1,4 +1,4 @@
-import EuropeanaApi from './apis/utils/base.js';
+import EuropeanaApi from './apis/base.js';
 import { BASE_URL as EUROPEANA_DATA_URL } from './data.js';
 import EuropeanaEntityApi from './entity.js';
