+++ packages/portal/src/plugins/europeana/fulltext.js
@@ -1,4 +1,4 @@
-import EuropeanaApi from './apis/utils/base.js';
+import EuropeanaApi from './apis/base.js';
 
 export default class EuropeanaFulltextApi extends EuropeanaApi {
   static ID = 'fulltext';
