+++ packages/portal/src/plugins/europeana/iiif/presentation.js
@@ -1,4 +1,4 @@
-import EuropeanaApi from '../apis/utils/base.js';
+import EuropeanaApi from '../apis/base.js';
 import { BASE_URL as EUROPEANA_IIIF_BASE_URL } from './index.js';
 
 export default class EuropeanaIiifPresentationApi extends EuropeanaApi {
