+++ packages/portal/src/plugins/europeana/media-proxy.js
@@ -1,4 +1,4 @@
-import EuropeanaApi from './apis/utils/base.js';
+import EuropeanaApi from './apis/base.js';
 
 export default class EuropeanaMediaProxyApi extends EuropeanaApi {
   static ID = 'mediaProxy';
