+++ packages/portal/src/plugins/europeana/recommendation.js
@@ -1,4 +1,4 @@
-import EuropeanaApi from './apis/utils/base.js';
+import EuropeanaApi from './apis/base.js';
 
 export default class EuropeanaRecommendationApi extends EuropeanaApi {
   static ID = 'recommendation';
