+++ packages/portal/src/plugins/europeana/thumbnail.js
@@ -1,6 +1,6 @@
 import md5 from 'md5';
 
-import EuropeanaApi from './apis/utils/base.js';
+import EuropeanaApi from './apis/base.js';
 
 export default class EuropeanaThumbnailApi extends EuropeanaApi {
   static ID = 'thumbnail';
